Fix peak fifo (#352)
* fix peak fifo

* fix mistake in spectrum.hpp
